section faction : modify controller in edit and add and modify manager with edit

+++ src/Controller/FactionController.php
@@ -53,9 +53,46 @@ class FactionController extends AbstractController
      */
     public function add() : string
     {
-      $nameError = null;
+      $nameError = $pictureError = $file_destination = null;
       if ($_SERVER["REQUEST_METHOD"] === "POST") {
         $isValid = true;
+
+        // download picture
+        if (!empty($_FILES['new-picture']['name']) && isset($_FILES['new-picture'])) {
+          $folder = 'faction';
+
+          $allowed = array('png', 'jpg', 'jpeg', 'gif');
+          $file_ext = explode('.', $_FILES['new-picture']['name']);
+          $file_ext = strtolower(end($file_ext));
+
+          $file_name_new = uniqid($folder.'-', false) . '.' . $file_ext;
+          $file_destination = 'assets/images/'.$folder.'/' . $file_name_new;
+
+          $filename = substr($_FILES['new-picture']['name'], 1);
+
+          if ($_FILES['new-picture']['size'] > 2097152) {
+            $pictureError = "La photo ne doit pas dépasser 2 Mo";
+            $isValid = false;
+          }
+          if(!in_array($file_ext, $allowed)) {
+            $pictureError = "La photo doit être au format jpg, jpeg, gif ou png";
+            $isValid = false;
+          }
+
+          if ($isValid) {
+            if (move_uploaded_file($_FILES['new-picture']['tmp_name'], $file_destination)) {
+              $_POST['picture'] = "/".$file_destination;
+              if (file_exists($filename)) {
+                unlink($filename);
+              }
+            } else {
+              $pictureError = "Erreur durant l'importation de la photo";
+              $isValid = false;
+            }
+          }
+        }
+        
+        // save data
         if (empty($_POST['name']) || !isset($_POST['name'])) {
           $nameError = "Merci de saisir un nom de faction";
           $isValid = false;
@@ -73,6 +110,7 @@ class FactionController extends AbstractController
 
       return $this->twig->render('Faction/add.html.twig', [
         'nameError' => $nameError,
+        'pictureError'  => $pictureError,
       ]);
     }
 
@@ -85,60 +123,58 @@ class FactionController extends AbstractController
      */
     public function edit(int $id) : string
     {
-      $nameError = $pictureError = null;
+      $nameError = $pictureError = $file_destination = null;
       if ($_SERVER["REQUEST_METHOD"] === "POST") {
         $isValid = true;
 
-        if (isset($_POST['modify-data'])) { // data
-
-          if (empty($_POST['name']) || !isset($_POST['name'])) {
-            $nameError = "Merci de saisir un nom de faction";
-            $isValid = false;
-          }
-          if ($isValid) {
-            if (empty($_POST['picture']) || !isset($_POST['picture'])) {
-              $_POST["picture"] = self::EMPTY_PICTURE;
-            }
-            $factionManager = new FactionManager();
-            $factionManager->editdataFaction($_POST, $id);
-            header('Location:/faction/edit/'.$id);
-          }
-
-        } else { // picture
-
+          // download picture
+        if (!empty($_FILES['new-picture']['name']) && isset($_FILES['new-picture'])) {
           $folder = 'faction';
 
-            $allowed = array('png', 'jpg', 'jpeg', 'gif');
-            $file_ext = explode('.', $_FILES['new-picture']['name']);
-            $file_ext = strtolower(end($file_ext));
+          $allowed = array('png', 'jpg', 'jpeg', 'gif');
+          $file_ext = explode('.', $_FILES['new-picture']['name']);
+          $file_ext = strtolower(end($file_ext));
 
-            $file_name_new = uniqid($folder.'-', false) . '.' . $file_ext;
-            $file_destination = 'assets/images/'.$folder.'/' . $file_name_new;
+          $file_name_new = uniqid($folder.'-', false) . '.' . $file_ext;
+          $file_destination = 'assets/images/'.$folder.'/' . $file_name_new;
 
-            $filename = substr($_POST['picture'], 1);
+          $filename = substr($_POST['picture'], 1);
 
-            if ($_FILES['new-picture']['size'] > 2097152) {
-              $pictureError = "La photo ne doit pas dépasser 2 Mo";
-              $isValid = false;
-            }
-            if(!in_array($file_ext, $allowed)) {
-              $pictureError = "La photo doit être au format jpg, jpeg, gif ou png";
-              $isValid = false;
-            }
+          if ($_FILES['new-picture']['size'] > 2097152) {
+            $pictureError = "La photo ne doit pas dépasser 2 Mo";
+            $isValid = false;
+          }
+          if(!in_array($file_ext, $allowed)) {
+            $pictureError = "La photo doit être au format jpg, jpeg, gif ou png";
+            $isValid = false;
+          }
 
-            if ($isValid) {
-              if (move_uploaded_file($_FILES['new-picture']['tmp_name'], $file_destination)) {
-                $factionManager = new FactionManager();
-                $factionManager->editPictureFaction(['picture' => '/'.$file_destination], $id);
-                if (file_exists($filename)) {
-                  unlink($filename);
-                }
-                header('Location:/faction/edit/'.$id);
-              } else {
-                $pictureError = "Erreur durant l'importation de la photo";
+          if ($isValid) {
+            if (move_uploaded_file($_FILES['new-picture']['tmp_name'], $file_destination)) {
+              $_POST['picture'] = "/".$file_destination;
+              if (file_exists($filename)) {
+                unlink($filename);
               }
+            } else {
+              $pictureError = "Erreur durant l'importation de la photo";
+              $isValid = false;
             }
+          }
+        }
+        
+        // save data
+        if (empty($_POST['name']) || !isset($_POST['name'])) {
+          $nameError = "Merci de saisir un nom de faction";
+          $isValid = false;
+        }
 
+        if ($isValid) {
+          if (empty($_POST['picture']) || !isset($_POST['picture'])) {
+            $_POST["picture"] = self::EMPTY_PICTURE;
+          }
+          $factionManager = new FactionManager();
+          $factionManager->editFaction($_POST, $id);
+          header('Location:/faction/list');
         }
       }
